Custom Lexer 적용
location 정보를 위하여 lexer 적용
lookahead는 1개로 처리

+++ zoker-parser/src/ast.rs
@@ -77,7 +77,7 @@ pub enum ExpressionType {
         arguments: Vec<Expression>,
     },
     Number {
-        value: i32,
+        value: u64,
     },
     Identifier {
         value: String,
